+++ Sources/SwiftUIKit/_Deprecated/Navigation/NavigationContext.swift
@@ -8,26 +8,21 @@
 
 import Foundation
 
-/**
- This context can be used to manage the state of a nav stack.
- */
+@available(*, deprecated, message: "This has been moved to https://github.com/danielsaidi/PresentationKit")
 public class NavigationContext<Screen: Hashable>: ObservableObject {
     
     public init() {}
 
     @Published
     public var path = [Screen]()
-}
-
-public extension NavigationContext {
 
     /// Pop a certain number of elements off the stack.
-    func goBack(_ steps: Int) {
+    public func goBack(_ steps: Int) {
         path.removeLast(steps)
     }
 
     /// Push a new screen onto the stack.
-    func push(_ screen: Screen) {
+    public func push(_ screen: Screen) {
         path.append(screen)
     }
 }