+++ RELEASE_NOTES.md
@@ -16,6 +16,9 @@ SwiftUIKit makes its best effort to honor semver, but breaking changes can occur
 * `FullScreenCoverContext` has been moved to https://github.com/danielsaidi/PresentationKit.
 * `Image+ScanCode` has been moved to https://github.com/danielsaidi/ScanCodes.
 * `LinkText` has been moved to https://github.com/danielsaidi/TextReplacements.
+* `NavigationButton` has been moved to https://github.com/danielsaidi/PresentationKit.
+* `NavigationLinkArrow` has been moved to https://github.com/danielsaidi/PresentationKit.
+* `NavigationContext` has been moved to https://github.com/danielsaidi/PresentationKit.
 * `PresentationContext` has been moved to https://github.com/danielsaidi/PresentationKit.
 * `ScanCodeType` has been moved to https://github.com/danielsaidi/ScanCodes.
 * `SheetContext` has been moved to https://github.com/danielsaidi/PresentationKit.